requirements.txt
Keras==2.0.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Pillow==8.1.0
Pillow-PIL==0.1.dev0
h5py==2.10.0
pip==21.0.1
scikit-image==0.18.1
scipy==1.1.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
